What's this about:
Add the code for Fuzzy Reasoning, using Mamdani and Centroid.

Expand to other territory of Fuzzy algorithm and method.

Expand to different language for language comparison.

Separate the docker file in order not to mess with the initial Docker configuration.

Modify the docker-compose.yml and DockerFile so it accepts Steel Bank
Common Lisp.

add supervisor.conf to called by DockerFile and docker-compose.yml

Separate the fuzzy logic into 2 directory which handle different
approaches.

create both Python and Common Lisp version of the code to handle main
file, with simple command due to it's only necessary for Github
checking.

Create the version of Common Lisp of the Python code above as a comparison,
minor fix on the Python and Common Lisp code (Need to double check the code later).

Hope that revising parts of the docker-build.yml file doesn't mess up the on web compilation.

Make license, due to the license used in Readme.md